+++ src/mako/application/cli/commands/app/GeneratePreloader.php
@@ -14,7 +14,7 @@ use mako\cli\input\Input;
 use mako\cli\output\Output;
 use mako\file\FileSystem;
 use mako\reactor\attributes\Arguments;
-use mako\reactor\attributes\Command as CommandAttribute;
+use mako\reactor\attributes\CommandDescription;
 use mako\reactor\Command;
 
 use function array_unique;
@@ -24,7 +24,7 @@ use function sprintf;
 /**
  * Command that generates an opcache preloader script.
  */
-#[CommandAttribute('app:generate-preloader', 'Generates a opcache preloader script.')]
+#[CommandDescription('Generates a opcache preloader script.')]
 #[Arguments(
 	new Argument('-i|--ignore-core-classes', 'Should the default selection of core classes be ignored?', Argument::IS_BOOL),
 	new Argument('-o|--output-path', 'Path to where the preloder script should be written', Argument::IS_OPTIONAL),