+++ src/mako/application/cli/commands/cache/Remove.php
@@ -10,12 +10,12 @@ namespace mako\application\cli\commands\cache;
 use mako\cache\CacheManager;
 use mako\cli\input\arguments\Argument;
 use mako\reactor\attributes\Arguments;
-use mako\reactor\attributes\Command as CommandAttribute;
+use mako\reactor\attributes\CommandDescription;
 
 /**
  * Command that removes the chosen key from the cache.
  */
-#[CommandAttribute('cache:remove', 'Removes the chosen key from the cache.')]
+#[CommandDescription('Removes the chosen key from the cache.')]
 #[Arguments(
 	new Argument('-c|--configuration', 'Configuration name', Argument::IS_OPTIONAL),
 	new Argument('-k|--key', 'Cache key'),