+++ src/mako/application/cli/commands/migrations/Up.php
@@ -9,12 +9,12 @@ namespace mako\application\cli\commands\migrations;
 
 use mako\cli\input\arguments\Argument;
 use mako\reactor\attributes\Arguments;
-use mako\reactor\attributes\Command as CommandAttribute;
+use mako\reactor\attributes\CommandDescription;
 
 /**
  * Command that runs all outstanding migrations.
  */
-#[CommandAttribute('migration:up', 'Runs all outstanding migrations.')]
+#[CommandDescription('Runs all outstanding migrations.')]
 #[Arguments(
 	new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
 )]