+++ CHANGELOG.md
@@ -43,7 +43,8 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Added `FileSystem::clearCache()` method.
 * Added `FileInfo::hasPermissions()` method that accepts an integer or a `Permissions` instance.
 * Added `FileInfo::getPermissions()` method that returns a `Permissions` instance.
-* Reactor command names and descriptions can now be registered with the `Command` attribute.
+* Reactor command names can now be registered with the `CommandName` attribute.
+* Reactor command descriptions can now be registered with the `CommandDescription` attribute.
 * Reactor command arguments can now be registered with the `Arguments` attribute.
 
 #### Changes
@@ -57,8 +58,8 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 
 #### Deprecations
 
-* Deprecated the `Command::$command` property. Use the `Command` attribute instead.
-* Deprecated the `Command::$description` property. Use the `Command` attribute instead.
+* Deprecated the `Command::$command` property. Use the `CommandName` attribute instead.
+* Deprecated the `Command::$description` property. Use the `CommandDescription` attribute instead.
 * Deprecated the `Command::getCommand()` method.
 * Deprecated the `Command::getDescription()` method.
 * Deprecated the `Command::getArguments()` method. Use the `Arguments` attribute instead.